CPSC 335 class Summer 2021
# Project-2

In this project we have implemented and compared two algorithms that solve the budgeted time-maximization problem. For one algorithm we use the greedy pattern. This algorithm always chooses the “best” (highest time-per-cost) ride item that fits within the budget and keeps selecting ride items in this manner until the budget is reached. For the second algorithm we use the exhaustive pattern, which is far slower that the greedy algorithm. Specifically, among all subsets of rides, the exhaustive algorithm returns the subset whose cost in dollars fits within the budget and whose time is the largest. Comparing both the algorithms and their time complexity, we elaborate and demonstrate the extent that the greedy algorithm outperforms the exhaustive algorithm.

Group members:

Sean Javiya - seanjaviya@csu.fullerton.edu

Shreeji Khalasi - shreejikhalasi@csu.fullerton.edu